disable initalizing new migration state accounts

+++ program/tests/start.rs
@@ -78,7 +78,7 @@ async fn start_migration() {
     let now = std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
         .unwrap()
-        .as_secs() as u64;
+        .as_secs();
 
     let mut state = migratorr.state().clone();
     state.status.unlock_time = now as i64 - 2;
@@ -451,7 +451,7 @@ async fn restart_migration() {
     let now = std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
         .unwrap()
-        .as_secs() as u64;
+        .as_secs();
 
     let mut state = migratorr.state().clone();
     state.status.unlock_time = now as i64 - 2;
